moved functions to utility files
